Associando pedido a transportadora

+++ src/main/java/br/com/techchalleng4/mslogistica/dto/CarrierDTO.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record CarrierDTO(
-        Long code,
+        Integer code,
         String name,
         Integer zipCodeStart,
         Integer zipCodeEnd
